V0.1
Corrected 404 error caused by incorrect routing, was due to renaming the pages folder. Despite updating relevant routing to reflect a change, it seems this needs to be left as "Pages" in order for the code behind the scenes to execute correctly.
Shared UI elements such as the navbar and titlebar have been created.